Initial commit of the SystemML deep learning experimental project.
This commit adds support for fully-connected (affine), feed-forward networks with ReLU nonlinearities, and L2 or L1 loss functions. The layers and loss functions all adopt a common forward/backward API.

Additionally, there is a "grad_check.dml" file that contain test functions to perform gradient checks on the computed gradients for the layers and loss functions.