Extend `util::pad_image` with a `pad_value` parameter
Currently, our `util::pad_image` function performs zero-padding,
which is correct for operations such as convolution.  In some cases,
such as max pooling, it is necessary to pad with a different value
such as negative infinity.  Therefore, this extends the
`util::pad_image` function to accept a `pad_value` parameter
with a noted typical value of 0.

Additionally, this fixes the max pooling layers by using
negative-infinity padding, and updates a test case accordingly.